requirements.txt
requests==2.24.0
instabot==0.117.0
pillow==7.0.0
python-dotenv==0.14.0